+++ web/components/interface/navigation/menu-dropdown/file-menu.tsx
@@ -1,14 +1,14 @@
-import { useMenuActions } from "@/lib/hooks/use-menu-actions";
-import { useRegisterMenuAction } from "@/lib/hooks/use-register-menu-action";
+import { useMenuActions } from "@/lib/hooks/menu-actions/use-menu-actions";
+import { useRegisterMenuAction } from "@/lib/hooks/menu-actions/use-register-menu-action";
 import { useTabViewManager } from "@/lib/hooks/use-tab-view-manager";
-import { ViewModeEnum } from "@pulse-editor/shared-utils";
 import { useEffect, useState } from "react";
 import { v4 } from "uuid";
 import NavMenuDropdown from "../nav-menu-dropdown";
 
 export default function FileMenuDropDown() {
   const { menuActions } = useMenuActions("file");
-  const { createTabView, activeTabView, closeTabView } = useTabViewManager();
+  const { createCanvasTabView, activeTabView, closeTabView } =
+    useTabViewManager();
 
   useRegisterMenuAction(
     {
@@ -20,7 +20,7 @@ export default function FileMenuDropDown() {
     },
     async () => {
       // Trigger new Workflow creation logic
-      await createTabView(ViewModeEnum.Canvas, { viewId: v4() });
+      await createCanvasTabView({ viewId: "canvas-" + v4() });
     },
     [],
   );