+++ web/components/views/canvas/canvas-view.tsx
@@ -1,12 +1,11 @@
+import { useRegisterMenuAction } from "@/lib/hooks/menu-actions/use-register-menu-action";
 import { useAppInfo } from "@/lib/hooks/use-app-info";
-import { useRegisterMenuAction } from "@/lib/hooks/use-register-menu-action";
-import useWorkflow from "@/lib/hooks/use-workflow";
+import useCanvasWorkflow from "@/lib/hooks/use-workflow";
 import {
   AppInfoModalContent,
   AppNodeData,
+  AppViewConfig,
   CanvasViewConfig,
-  MenuAction,
-  Workflow,
 } from "@/lib/types";
 import { Button } from "@heroui/react";
 import { Action } from "@pulse-editor/shared-utils";
@@ -23,12 +22,11 @@ import {
   Edge as ReactFlowEdge,
   Node as ReactFlowNode,
   reconnectEdge,
-  useNodes,
   useReactFlow,
   useViewport,
 } from "@xyflow/react";
 import "@xyflow/react/dist/style.css";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useRef } from "react";
 import Icon from "../../misc/icon";
 import AppNode from "./nodes/app-node/app-node";
 import "./theme.css";
@@ -51,69 +49,46 @@ Pulse Editor is a modular, cross-platform, AI-powered creativity platform with f
 `,
 };
 
-export default function CanvasView({ config }: { config?: CanvasViewConfig }) {
+export default function CanvasView({ config }: { config: CanvasViewConfig }) {
   const { openAppInfoModal } = useAppInfo();
 
-  const [workflow, setWorkflow] = useState<Workflow | undefined>(undefined);
-  const [entryPoint, setEntryPoint] = useState<
-    ReactFlowNode<AppNodeData> | undefined
-  >(undefined);
-
-  const { startWorkflow, runningNodes } = useWorkflow(workflow, entryPoint);
+  const {
+    workflow,
+    entryPoint,
+    startWorkflow,
+    updateWorkflowEdges,
+    updateWorkflowNodes,
+    exportWorkflow,
+    updateWorkflowNodeData,
+  } = useCanvasWorkflow(config.viewId);
 
   const viewport = useViewport();
   const { screenToFlowPosition } = useReactFlow();
-  const nodes = useNodes<ReactFlowNode<AppNodeData>>();
 
   const containerRef = useRef<HTMLDivElement>(null);
 
+  // this is called when new node is added below using updateWorkflowNodes
   const onNodesChange = useCallback(
-    (changes: NodeChange<ReactFlowNode<AppNodeData>>[]) => {
-      console.log("Node changes:", changes);
-      setWorkflow((prev) => {
-        if (!prev) return undefined;
-        return {
-          ...prev,
-          nodes: applyNodeChanges(changes, prev.nodes),
-        };
-      });
-    },
-    [],
+    (changes: NodeChange<ReactFlowNode<AppNodeData>>[]) =>
+      updateWorkflowNodes((oldNodes) => applyNodeChanges(changes, oldNodes)),
+    [updateWorkflowNodes],
   );
   const onEdgesChange = useCallback(
-    (changes: EdgeChange<{ id: string; source: string; target: string }>[]) => {
-      console.log("Edge changes:", changes);
-      setWorkflow((prev) => {
-        if (!prev) return undefined;
-        return {
-          ...prev,
-          edges: applyEdgeChanges(changes, prev.edges),
-        };
-      });
-    },
-    [],
+    (changes: EdgeChange<{ id: string; source: string; target: string }>[]) =>
+      updateWorkflowEdges((oldEdges) => applyEdgeChanges(changes, oldEdges)),
+    [updateWorkflowEdges],
   );
   const onConnect = useCallback(
     (params: any) =>
-      setWorkflow((prev) => {
-        if (!prev) return undefined;
-        return {
-          ...prev,
-          edges: addEdge(params, prev.edges),
-        };
-      }),
-    [],
+      updateWorkflowEdges((oldEdges) => addEdge(params, oldEdges)),
+    [updateWorkflowEdges],
   );
   const onReconnect = useCallback(
     (oldEdge: ReactFlowEdge, newConnection: Connection) =>
-      setWorkflow((prev) => {
-        if (!prev) return undefined;
-        return {
-          ...prev,
-          edges: reconnectEdge(oldEdge, newConnection, prev.edges),
-        };
-      }),
-    [],
+      updateWorkflowEdges((oldEdges) =>
+        reconnectEdge(oldEdge, newConnection, oldEdges),
+      ),
+    [updateWorkflowEdges],
   );
 
   /* Node creator functions */
@@ -132,43 +107,8 @@ export default function CanvasView({ config }: { config?: CanvasViewConfig }) {
       icon: "download",
     },
 
-    async () => {
-      await exportWorkflow();
-    },
-    [workflow],
-  );
-  useRegisterMenuAction(
-    {
-      name: "Import Workflow",
-      menuCategory: "file",
-      description: "Import a workflow from a JSON file",
-      shortcut: "Ctrl+Alt+I",
-      icon: "upload",
-    },
-    async () => {
-      const input = document.createElement("input");
-      input.type = "file";
-      input.accept = "application/json";
-      input.onchange = (e: any) => {
-        const file = e.target.files[0];
-        const reader = new FileReader();
-        reader.onload = (event) => {
-          try {
-            const workflow = JSON.parse(event.target?.result as string);
-            if (workflow) {
-              setWorkflow(workflow);
-            } else {
-              alert("Invalid workflow file");
-            }
-          } catch (err) {
-            alert("Error reading workflow file");
-          }
-        };
-        reader.readAsText(file);
-      };
-      input.click();
-    },
-    [],
+    async () => exportWorkflow(),
+    [exportWorkflow],
   );
   useRegisterMenuAction(
     {
@@ -185,139 +125,100 @@ export default function CanvasView({ config }: { config?: CanvasViewConfig }) {
     [entryPoint],
   );
 
-  // Add or remove nodes when config changes
+  // Promote nodes to workflow nodes,
+  // or remove workflow nodes that are no longer in the config
   useEffect(() => {
+    function promoteToWorkflowNode(newApps: AppViewConfig[]) {
+      const newNodes: ReactFlowNode<AppNodeData>[] =
+        newApps?.map((appConfig) => {
+          const flowCenter = getViewCenter(appConfig);
+
+          const newAppNodeData: AppNodeData = {
+            config: appConfig,
+            selectedAction: undefined,
+            setSelectedAction: async (action: Action | undefined) => {
+              await updateWorkflowNodeData(appConfig.viewId, {
+                selectedAction: action,
+              });
+            },
+            isRunning: false,
+          };
+
+          return {
+            id: appConfig.viewId,
+            position: flowCenter,
+            data: newAppNodeData,
+            type: "appNode",
+            height: appConfig.recommendedHeight ?? 360,
+            width: appConfig.recommendedWidth ?? 640,
+          };
+        }) ?? [];
+
+      // Add new nodes to the workflow
+      updateWorkflowNodes((oldNodes) => oldNodes.concat(newNodes));
+    }
+
+    function removeWorkflowNode(removedNodes: ReactFlowNode<AppNodeData>[]) {
+      updateWorkflowNodes((oldNodes) => {
+        return oldNodes.filter(
+          (node) =>
+            !removedNodes.find((removedNode) => removedNode.id === node.id),
+        );
+      });
+    }
+
+    function getViewCenter(appConfig: AppViewConfig) {
+      const containerBounds = containerRef.current?.getBoundingClientRect();
+
+      if (!containerBounds) throw new Error("Container bounds not found");
+
+      const screenCenter = {
+        x:
+          containerBounds.left +
+          containerBounds.width / 2 -
+          ((appConfig.recommendedWidth ?? 640) / 2) * viewport.zoom,
+        y:
+          containerBounds.top +
+          containerBounds.height / 2 -
+          ((appConfig.recommendedHeight ?? 360) / 2) * viewport.zoom,
+      };
+
+      const flowCenter = screenToFlowPosition(screenCenter);
+      return flowCenter;
+    }
+
     if (config) {
-      // Added nodes
-      const newNodes = config.nodes?.filter(
+      // Added apps
+      const addedApps = config.appConfigs?.filter(
         (newNode) =>
           !workflow?.nodes.find((node) => node.id === newNode.viewId),
       );
 
-      if (newNodes && newNodes.length > 0) {
-        const newAppNodes: ReactFlowNode<AppNodeData>[] =
-          newNodes?.map((appConfig) => {
-            const containerBounds =
-              containerRef.current?.getBoundingClientRect();
-
-            if (!containerBounds) throw new Error("Container bounds not found");
-
-            const screenCenter = {
-              x:
-                containerBounds.left +
-                containerBounds.width / 2 -
-                ((appConfig.recommendedWidth ?? 640) / 2) * viewport.zoom,
-              y:
-                containerBounds.top +
-                containerBounds.height / 2 -
-                ((appConfig.recommendedHeight ?? 360) / 2) * viewport.zoom,
-            };
-
-            const flowCenter = screenToFlowPosition(screenCenter);
-
-            const appNodeData: AppNodeData = {
-              config: appConfig,
-              selectedAction: undefined,
-              setSelectedAction: async (action: Action | undefined) => {
-                setWorkflow((prev) => {
-                  if (!prev) return undefined;
-                  return {
-                    ...prev,
-                    nodes: prev.nodes.map((node) => {
-                      if (node.id === appConfig.viewId) {
-                        return {
-                          ...node,
-                          data: {
-                            ...node.data,
-                            selectedAction: action,
-                          },
-                        };
-                      }
-                      return node;
-                    }),
-                  };
-                });
-              },
-              isRunning:
-                runningNodes?.find((n) => n.id === appConfig.viewId) !==
-                undefined,
-            };
-
-            return {
-              id: appConfig.viewId,
-              position: flowCenter,
-              data: appNodeData,
-              type: "appNode",
-              height: appConfig.recommendedHeight ?? 360,
-              width: appConfig.recommendedWidth ?? 640,
-            };
-          }) ?? [];
-
-        // Add new nodes to the workflow
-        setWorkflow((prev) => {
-          if (!prev) {
-            return {
-              nodes: newAppNodes,
-              edges: [],
-            };
-          }
-          return {
-            ...prev,
-            nodes: prev.nodes.concat(newAppNodes),
-          };
-        });
+      if (addedApps && addedApps.length > 0) {
+        promoteToWorkflowNode(addedApps);
       }
 
-      // Removed nodes
-      const removedNodes = workflow?.nodes.filter(
-        (node) => !config.nodes?.find((newNode) => newNode.viewId === node.id),
+      // Removed apps
+      const removedApps = workflow?.nodes.filter(
+        (node) =>
+          !config.appConfigs?.find((newNode) => newNode.viewId === node.id),
       );
 
-      if (removedNodes && removedNodes.length > 0) {
-        setWorkflow((prev) => {
-          if (!prev) return undefined;
-          return {
-            ...prev,
-            nodes: prev.nodes.filter(
-              (node) =>
-                !removedNodes.find((removedNode) => removedNode.id === node.id),
-            ),
-          };
-        });
+      if (removedApps && removedApps.length > 0) {
+        removeWorkflowNode(removedApps);
       }
     }
-  }, [config, viewport, runningNodes]);
-
-  useEffect(() => {
-    const selectedNodes = nodes.filter((node) => node.selected);
-    if (selectedNodes.length > 0) {
-      setEntryPoint(selectedNodes[0]);
-    } else {
-      setEntryPoint(undefined);
-    }
-  }, [nodes]);
-
-  async function exportWorkflow() {
-    const blob = new Blob([JSON.stringify(workflow, null, 2)], {
-      type: "application/json",
-    });
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement("a");
-    a.href = url;
-    a.download = "workflow.json";
-    a.click();
-    URL.revokeObjectURL(url);
-  }
+  }, [config, viewport, screenToFlowPosition, updateWorkflowNodes]);
 
   return (
     <div
       ref={containerRef}
       className="bg-content3 text-content3-foreground relative h-full w-full"
-      id={`canvas-${config?.viewId}`}
+      id={config.viewId}
     >
       <ReactFlow
-        nodes={workflow?.nodes}
-        edges={workflow?.edges}
+        nodes={workflow?.nodes ?? []}
+        edges={workflow?.edges ?? []}
         onNodesChange={onNodesChange}
         onEdgesChange={onEdgesChange}
         onConnect={onConnect}
@@ -338,7 +239,7 @@ export default function CanvasView({ config }: { config?: CanvasViewConfig }) {
           },
         }}
       >
-        <Background variant={BackgroundVariant.Dots} />
+        <Background id={config.viewId} variant={BackgroundVariant.Dots} />
       </ReactFlow>
       <Button
         isIconOnly