+++ web/components/views/canvas/nodes/app-node/app-node.tsx
@@ -1,7 +1,6 @@
 import useScopedActions from "@/lib/hooks/use-scoped-actions";
 import { useTabViewManager } from "@/lib/hooks/use-tab-view-manager";
 import { AppNodeData } from "@/lib/types";
-import { ViewModeEnum } from "@pulse-editor/shared-utils";
 import { Node } from "@xyflow/react";
 import { memo } from "react";
 import { v4 } from "uuid";
@@ -15,13 +14,13 @@ const AppNode = memo((props: any) => {
     nodeProps.data;
   const viewId = config.viewId;
 
-  const { createTabView, deleteAppViewInCanvasView } = useTabViewManager();
+  const { createAppTabView, deleteAppViewInCanvasView } = useTabViewManager();
   const { actions } = useScopedActions(config.app);
 
   async function openViewInFullScreen() {
-    await createTabView(ViewModeEnum.App, {
+    await createAppTabView({
       ...config,
-      viewId: v4(),
+      viewId: `${config.app}-${v4()}`,
     });
   }