+++ web/components/views/canvas/nodes/backend-node/backend-node.tsx
@@ -2,7 +2,6 @@ import BaseAppView from "@/components/views/base/base-app-view";
 import useScopedActions from "@/lib/hooks/use-scoped-actions";
 import { useTabViewManager } from "@/lib/hooks/use-tab-view-manager";
 import { AppNodeData } from "@/lib/types";
-import { ViewModeEnum } from "@pulse-editor/shared-utils";
 import { Node } from "@xyflow/react";
 import { memo } from "react";
 import { v4 } from "uuid";
@@ -16,13 +15,13 @@ const BackendNode = memo((props: any) => {
     nodeProps.data;
   const viewId = config.viewId;
 
-  const { createTabView, deleteAppViewInCanvasView } = useTabViewManager();
+  const { createAppTabView, deleteAppViewInCanvasView } = useTabViewManager();
   const { actions } = useScopedActions(config.app);
 
   async function openViewInFullScreen() {
-    await createTabView(ViewModeEnum.App, {
+    await createAppTabView({
       ...config,
-      viewId: v4(),
+      viewId: `${config.app}-${v4()}`,
     });
   }